211021
button - now the button is inactive if the value is empty. Combed my hair.
cards - the cards are not displayed until the recipes appear.
ingredients - the key is now equal to the ingredient, since the ingredients cannot match - everything will be fine.
search - the local state has been removed, now input has become a global variable. Added an action on "enter" - adds an ingredient to the list.
upbar&about - everything became English.
spoonReducer, types - the types now correspond to the action.
spoonState - added a check that the ingredient is already in the ingredients.