21.10.22
button renamed to buttonAddIngredient.
buttonAddIngredient - the button now changes color if the ingredient cannot be added.
buttonFindRecipe - added a new button to search for recipes.
ingredient - the ingredient is highlighted if you are trying to add an existing ingredient. Button to remove an ingredient from the ingredient lists has started working.
search - addIngredient2Ingredients now passes the value to add to the ingredients list.
spoonReducer, spoonState, types - added the function of removing an ingredient from the list of ingredients.