21.11.03
analyzedInstructions.js
The component is responsible for step-by-step cooking instructions.

extendedIngredients.js
Added rounding for ingredients with ml.

fullRecipe.js
Added useRef to scroll the screen to the step-by-step instructions.
Added the "Step-by-step instructions" button to enable step-by-step instructions.
Added information panel for summary.
Added a panel with all the ingredients for instructions.

summaryUpBar.js
The list of all ingredients is now a separate panel.